pyproject.toml
[project]
name = "dyn-advisor"
version = "0.1.0"
description = "A small, inspectable agent that indexes a Dynamo graph repository and recommends graphs based on user intent"
authors = [
    {name = "DaltonGOO"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
dyn-advisor = "dyn_advisor.cli:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dyn_advisor*"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
